Stabilize server:info tests (#160)
The server:info API route returns a very long string, that sometimes trigger a linebreak whence the stdout buffer is full. This makes the test regexp fail as it tests the response as a single line.

This PR makes the regexp less strict, only testing the beginning of the response: we can safely assume that the string is a well-formed JSON object because, otherwise, the conversion to a JSON object would trigger an exception.