Fix the documentation rebuild time in development mode (#193)
Since the Javascript bundling has been moved to webpack, documentation rebuilds were very slow.
This is due to the uglifier plugin, which is very useful in production to reduce the size of the bundle, but makes the documentation hot reload a burden in dev mode.

This PR deactives the uglifier plugin in dev mode. This reduce the rebuild time from 6-9s on my machine to 150-300ms.