auth:checkToken: remove an erroneous and confusing precision (#274)
# Description

In our `auth:checkToken` documentation there is the following sentence: `This API route does not require the caller to be logged in.`

After thinking hard about it I _think_ I understand what it means: "you don't need to have an authentication token active to verify a token". Which should be obvious, since we require the token to be checked to be set in the body part of the request.

Moreover, even with a more explicit comment, that would still be confusing, because Kuzzle doesn't give any special treatment to this API route: administrators can remove rights to it to the anonymous role, meaning that you cannot use it without being logged in (or at all). That may sounds weird, but that's how it is.

# Other changes

- The CI was still referring to the `documentation-v2` repository name, so I fixed that.
- The lint check wasn't executed in the CI, which is a mistake
- Fixed a linter error to make the CI pass