Make security guides more readable (#275)
# Description

1. The security guide features examples using Javascript code instead of plain JSON objects, and this was confusing for our readers
2. Make the security guide examples clearer
3. Wrap the very, very long `curl` example to prevent an impractical horizontal scrollbar from appearing (the example still works after having been copy-pasted, obviously)

# How to test

Go to the following pages in the netlify preview:

- Security guide: guide/1/essentials/security ([current production version](https://docs.kuzzle.io/guide/1/essentials/security))
- User authentication guide: guide/1/essentials/user-authentication/ ([current production version](https://docs.kuzzle.io/guide/1/essentials/user-authentication/))

# Note

~There is a handlebar pattern that is not interpreted at the bottom of the security guide (this is already the case in production). 
I don't have much time to spend on that, but AFAIK it seems that handlebars helpers returning a safestring prevent automatic field interpolation appearing later on the same line. To be investigated.~

Fixed two of those actually. The problem is that, for still unknown reasons, using a handlebars safestring disables further markdown interpretation. So, until we understand what's going on, I used HTML instead of markdown.