[fix] Fixed height of TOC (#352)
The height of the TOC (on the right side of the page) was set to 100% by default, but this overlapped with the social links on the footer. The height of the TOC is now set to the same value of the height of the Sidebar. The social links are now clickable.