[algolia] Record size, partial indexation, same-branch framework (#418)
#### Limit the size of indexation objects to respect the Algolia quota
Adds the logic to limit the content of a record, after the errors reported on [this Travis build](https://travis-ci.org/kuzzleio/documentation/jobs/626583427#L333). Algolia explicitly [discourages creating big records](https://www.algolia.com/doc/faq/basics/is-there-a-size-limit-for-my-index-records/), like we do for the [page of the error-codes](https://docs.kuzzle.io/core/2/api/essentials/errors/codes/).
In the mid-term, we should implement a logic that allows to split a page in many different records if its contents are too heavy for Algolia. For the time-being, we just chunk the content, which is obviously a shitty solution.

#### Same-branch framework
Work has been done to make it possible to build the sub-repositories using the framework from the same branch. I changed the `.repos/repositories.js` in order to link the `src/.vuepress` code from the root-repository, right into the sub-repository tree, and invoke VuePress from the root-repository tree. This way, we are progressively getting out of the way the `<sub-repo>/doc/doc.sh` script (which duplicates logic from the `repositories.js` script) and we are avoiding to repeatedly clone the root-repository from the sub-repositories.

#### Selective deletion of the pre-indexed objects
Sub-repositories now push index objects with a special `_tags` attribute, that contains the repository name. This is useful to delete these objects before re-indexing each sub-repository.