Indexation to Algolia (#419)
* [algolia] Fixed indexation (#417)

## What does this PR do?
Updated the Algolia API keys  to use the new indices of the support@kuzzle.io account. This 

### How should this be manually tested?
1. Wait for the CI to deploy the docs
2. Search something
3. Make sure the results are meaningful and the links aren't broken

### Other changes
Written some proper explanation of how this whole octopus works.

### Boyscout
Added an error message in case the search fails.

* [algolia] Record size, partial indexation, same-branch framework (#418)


#### Limit the size of indexation objects to respect the Algolia quota
Adds the logic to limit the content of a record, after the errors reported on [this Travis build](https://travis-ci.org/kuzzleio/documentation/jobs/626583427#L333). Algolia explicitly [discourages creating big records](https://www.algolia.com/doc/faq/basics/is-there-a-size-limit-for-my-index-records/), like we do for the [page of the error-codes](https://docs.kuzzle.io/core/2/api/essentials/errors/codes/).
In the mid-term, we should implement a logic that allows to split a page in many different records if its contents are too heavy for Algolia. For the time-being, we just chunk the content, which is obviously a shitty solution.

#### Same-branch framework
Work has been done to make it possible to build the sub-repositories using the framework from the same branch. I changed the `.repos/repositories.js` in order to link the `src/.vuepress` code from the root-repository, right into the sub-repository tree, and invoke VuePress from the root-repository tree. This way, we are progressively getting out of the way the `<sub-repo>/doc/doc.sh` script (which duplicates logic from the `repositories.js` script) and we are avoiding to repeatedly clone the root-repository from the sub-repositories.

#### Selective deletion of the pre-indexed objects
Sub-repositories now push index objects with a special `_tags` attribute, that contains the repository name. This is useful to delete these objects before re-indexing each sub-repository.

* [ci] using new repo scripts in .travis.yml

* [ci] fixes to indexation
* Not clearing the index in the framework job, as we cannot foresee the completion order.
* not hardcoding `/doc/` in prepareRepository, as GO doc resides in `.doc`
* Fixed error in the path of the JS SDK 7 doc

* [ci] simplified sdk sub-repos paths

* [ci] travis.yml
* Fixed cloudfront cache-invalidation for sub-repos
* Removed encrypted Algolia Write Key

* [ci] fixed typo in .travis.yml

* [ci] I'm tired

* [ci] misc
* show sdout of executeCommand
* simplify clone path of plugin sub repos

* [ci] run command output

* [ci] add SITE_BASE to buildRepository script

* [ci] add / to base_url

* [algolia] using ctx.base as basePath

* [search] ugli ack to avoid double slashes in result URLs

* [misc]
* Added missing V2 plugins in .travis.yml
* Renamed plugins
* Tiles dynamic height
* Changed symlinks destinations according to repositories.yml

* [search] fixed duplicate keys

* [misc]
* fix resolution of valid first child on root nodes
* added prettierrc

* [sidebar] getPageChildre

* [root] make root doc a page

* [git] Deleted accidentally added submodule

* [snippets] fix paths (hope so)

* [search] fix go to highlighted result

* Updated Readme

* Updated Readme

* Removed accidentally versioned subrepo

* Better major version selector (#420)

Add a new kuzzle major version selector on header

* [css] fix major version selector height

* [hotfix] better selector width & fix default major in number

Co-authored-by: Nicolas Juelle <n.juelle@gmail.com>